Struct threshold_crypto::poly::Poly

pub struct Poly { /* fields omitted */ }

A univariate polynomial in the prime field.

Implementations

impl Poly

pub fn random<R: Rng>(degree: usize, rng: &mut R) -> Self

Creates a random polynomial.

Panics

Panics if the degree is too large for the coefficients to fit into a Vec.

pub fn try_random<R: Rng>(degree: usize, rng: &mut R) -> Result<Self>

Creates a random polynomial. This constructor is identical to the Poly::random() constructor in every way except that this constructor will return an Err where try_random would return an error.

pub fn zero() -> Self

Returns the polynomial with constant value 0.

pub fn is_zero(&self) -> bool

Returns true if the polynomial is the constant value 0.

pub fn one() -> Self

Returns the polynomial with constant value 1.

pub fn constant(c: Fr) -> Self

Returns the polynomial with constant value c.

pub fn identity() -> Self

Returns the identity function, i.e. the polynomial “x”.

pub fn monomial(degree: usize) -> Self

Returns the (monic) monomial: x.pow(degree).

pub fn interpolate<T, U, I>(samples_repr: I) -> Self where
I: IntoIterator<Item = (T, U)>,
T: IntoFr,
U: IntoFr,

Returns the unique polynomial f of degree samples.len() - 1 with the given values (x, f(x)).

pub fn degree(&self) -> usize

Returns the degree.

pub fn evaluate<T: IntoFr>(&self, i: T) -> Fr

Returns the value at the point i.

pub fn commitment(&self) -> Commitment

Returns the corresponding commitment.

pub fn reveal(&self) -> String

Generates a non-redacted debug string. This method differs from the Debug implementation in that it does leak the secret prime field elements.